requirements.txt
functions-framework==3.*
qdrant-client
cohere
python-dotenv 

